requirements.txt
argparse
future
biopython
pandas
docopt
schema
pyvcf
sh
funcy
tabulate
